chapters/chapter14/requirements.txt
apache-airflow[aws]~=1.10.11
minio~=5.0.10
